+++ js_modules/dagit/packages/core/src/app/Util.tsx
@@ -20,6 +20,10 @@ export const formatElapsedTime = (msec: number) => {
   return `${hours}:${twoDigit(min)}:${twoDigit(sec)}`;
 };
 
+export function displayNameForAssetKey(key: {path: string[]}) {
+  return key.path.join(' > ');
+}
+
 export function breakOnUnderscores(str: string) {
   return str.replace(/_/g, '_\u200b');
 }
