+++ js_modules/dagit/packages/core/src/assets/AssetNeighborsGraph.tsx
@@ -6,12 +6,7 @@ import {SVGViewport} from '../graph/SVGViewport';
 import {AssetLinks} from '../workspace/asset-graph/AssetLinks';
 import {AssetNode} from '../workspace/asset-graph/AssetNode';
 import {ForeignNode} from '../workspace/asset-graph/ForeignNode';
-import {
-  layoutGraph,
-  GraphData,
-  displayNameForAssetKey,
-  LiveData,
-} from '../workspace/asset-graph/Utils';
+import {layoutGraph, GraphData, LiveData} from '../workspace/asset-graph/Utils';
 import {RepoAddress} from '../workspace/types';
 
 import {AssetNodeDefinitionFragment} from './types/AssetNodeDefinitionFragment';
